feat(#192): Calendar Select 페이지 내 UI 분리, Month 공통 컴포넌트 생성
