+++ src/pages/mytrip/DatesSelectPage/style.ts
@@ -1,6 +1,11 @@
 import styled from 'styled-components';
 
 export const HeadingWrapper = styled.div`
-  margin-top: 16px;
-  margin-bottom: 42px;
+  padding-top: 15px;
+  margin-bottom: 15px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 15px;
 `;
